Add updates for layout
