--- index.js
@@ -1,18 +1,0 @@
-
-import BorderLayout from './lib/BorderLayout';
-import Center from './lib/Center';
-import Footer from './lib/Footer';
-import Header from './lib/Header';
-import HorizontalLinearLayout from './lib/HorizontalLinearLayout';
-import LinearLayout from './lib/LinearLayout';
-import VerticalLinearLayout from './lib/VerticalLinearLayout';
-
-export default {
-  BorderLayout,
-  Center,
-  Footer,
-  Header,
-  HorizontalLinearLayout,
-  LinearLayout,
-  VerticalLinearLayout
-};
