+++ layout.js
@@ -1,9 +1,91 @@
-/*eslint no-unused-vars: 0*/
-
 import React, { Component } from 'react';
 import { View } from 'react-native';
 
-import Overlay from './Overlay';
+export const orientation = {
+  horizontal: 'horizontal',
+  vertical: 'vertical'
+};
+
+export default class LinearLayout extends Component {
+  render() {
+    return (
+      <View style={ { flexDirection: this.props.orientation == orientation.horizontal ? 'row' : 'column' } }>
+        { this.props.children }
+      </View>
+      );
+  }
+}
+
+export default class VerticalLinearLayout extends Component {
+  render() {
+    return (
+      <View style={ { flexDirection: 'column' } }>
+        { this.props.children }
+      </View>
+      );
+  }
+}
+
+
+export default class Overlay extends Component {
+  render() {
+    return (
+      <View style={ [{ position: 'absolute', top: 0, left: 0, right: 0, bottom: 0, backgroundColor: 'transparent' }, this.props.style] }>
+        { this.props.children }
+      </View>
+      );
+  }
+}
+
+export default class HorizontalLinearLayout extends Component {
+  render() {
+    return (
+      <View style={ { flexDirection: 'row' } }>
+        { this.props.children }
+      </View>
+      );
+  }
+}
+
+export class Header extends Component {
+  render() {
+    return (
+      <View style={ { position: 'absolute', top: 0, left: 0, right: 0 } }>
+        { this.props.children }
+      </View>
+      );
+  }
+}
+
+export class Footer extends Component {
+  render() {
+    return (
+      <View style={ { position: 'absolute', left: 0, right: 0, bottom: 0 } }>
+        { this.props.children }
+      </View>
+      );
+  }
+}
+
+export class Fill extends Component {
+  render() {
+    return (
+      <View style={ [{ flex: 1 }, this.props.style] }>
+        { this.props.children }
+      </View>
+      );
+  }
+}
+
+export class Center extends Component {
+  render() {
+    return (
+      <View style={ [{ flex: 1, alignItems: 'center', justifyContent: 'center' }, this.props.style] }>
+        { this.props.children }
+      </View>
+      );
+  }
+}
 
 const reorderSubviews = subviews => {
   if (!subviews.findIndex) {
@@ -38,9 +120,11 @@ const reorderSubviews = subviews => {
 
     const reordered = skipped.concat(before).concat(reorderSubviews(content)).concat(after).concat(overlays);
 
-    return <View style={ { flex: 1, flexDirection: isHorizontal ? 'row' : 'column' } }>
-             { reordered }
-           </View>;
+    return (
+      <View style={ { flex: 1, flexDirection: isHorizontal ? 'row' : 'column' } }>
+        { reordered }
+      </View>
+      );
   }
 
   return subviews;
@@ -74,12 +158,3 @@ export class Top extends Layout {
 }
 export class Bottom extends Layout {
 }
-
-export default {
-  Layout,
-  FillLayout,
-  Top,
-  Left,
-  Right,
-  Bottom
-};
