--- lib/Footer.js
@@ -1,14 +1,0 @@
-/*eslint no-unused-vars: 0*/
-
-import React, { Component } from 'react';
-import { View } from 'react-native';
-
-export default class Footer extends Component {
-  render() {
-    return (
-      <View style={ { position: 'absolute', left: 0, right: 0, bottom: 0 } }>
-        { this.props.children }
-      </View>
-      );
-  }
-}
